Add FIP64 configuration API
This patch allows the user to configure the mappings between ip6<->ip4
addresses. You can see the commands on the vpp cli by calling help:

(vppcli)$ fip64 help
  add                            <ip6 address> <ip4 address>
  del                            <ip6 address>

The configuration stores the mapping in an in-memory hashmap so the
ip6-fip64 and ip4-fip64 nodes can query the mappings at will.
Adding/deleting a mapping also adds/deletes the corresponding
adjacencies on the routing tables of the ip4/ip6 lookup nodes.

Change-Id: Iec18798cbe102a67fab2cd26576e35b17c474503
Signed-off-by: Xavi León <xavi.leon@midokura.com>
Signed-off-by: Miguel Herranz <miguel@midokura.com>