Add unit testing "framework" for fip64
It doesn't look like we'll be able to unit test the node itself. There's
too much wiring needed. But we can unit tests the other bits.

This patch enables the building of the tests. See vnet/test/README for
instructions to run them.

This patch also refactors some methods to be able to unit test the
mapping functions. We should aim to keep the rest of our code testable
like this. Avoid globals if at all possible, and only pass the
vlib_main_t as a parameter when absolutely necessary.

Change-Id: I91016710707d1bb146fe02a100e564104acc9191
Signed-off-by: Ivan Kelly <ivan@midokura.com>