FIP64: Move mapping bi-hash from fip to tenant
We were keeping mappings for every fip, which was silly as the
address pool is shared between all fips on a given tenant. With
this patch the mappings are kept in the tenant structure.

Now an external IPv6 client will be mapped to the same IPv4
address when communicating with different FIPs on the same
tenant.

Change-Id: I70ce8d2f2518512a335f1574fff7cf6b26b7b36d
Signed-off-by: Adrian Serrano <adrian@midokura.com>