Update patch set 1
Patch Set 1:

This route is not for flowstate packets, but for regular outgoing packets which need to pass through ip4-fip64 in order to be translated to ip6.

Currently we're doing this by adding a route on every tenant table:
pool_address/pool_prefix via ip4-fip64

But once we have flow state, we might eventually receive packets from different pools (i.e. when another vpp node goes down) so we need the default route.

Patch-set: 1