Fip64: Fix crash dealing with table ids.
When setting the sw_if_index field of a vnet buffer, we were using the
table id, but a table index is required instead. Table ids are symbolic
while table indexes are sequential in order of creation. This caused a
crash when using a table id that wasn't allocated sequentially.

Change-Id: I51a8324c1e6c8535f07098ac8d6de58eba23689d
Signed-off-by: Adrian Serrano <adrian@midokura.com>