VPP binary API for FIP64
This change add a VPP binary API for FIP64, with methods:

  fip64_add
  fip64_del
  fip64_sync_enable
  fip64_sync_disable

It also adds the needed code and help descriptions to allow them to
be used in VPP API tester (make build-vat && make run-vat).

Finally, new options has been added to easily test the methods
with the 'test_client' program.

Change-Id: I4957746891831d1eb9417a8bd6b657333a2bf79c
Signed-off-by: Miguel Herranz <miguel@midokura.com>