Make version and deps version for rpm consistent
The rpm spec previously used a %define for the version information,
which was in turn generated using a script. However, %define is lazily
evaluated, so if the version script returned different things for two
separate invocations, the vpp package could depend on a version of the
vpp-lib package that does not exist. This can happen if there's a
timestamp element in the version, and the invocation for the packages
version occurs in a different second to the invocation for the
dependencies version.

%global solves this problem as it is expanded only once, at definition
 time.

Change-Id: Icba4a150c4f70d6c4bf5d311dcfd5812c581e843
Signed-off-by: Ivan Kelly <ivan@midokura.com>